backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.5.1
flask-jwt-extended==4.7.1
bcrypt==4.3.0
python-dotenv==1.1.0
eventlet==0.37.0
